Further refactor middlewares and viewers
