+++ lib/app_profiler/viewer/firefox.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require "app_profiler/viewer/middleware/speedscope"
+require "app_profiler/viewer/middleware/firefox"
 
 module AppProfiler
   module Viewer
-    class SpeedscopeRemoteViewer < BaseViewer
+    class FirefoxViewer < BaseViewer
       class << self
         def view(profile, params = {})
           new(profile).view(**params)
@@ -17,10 +17,10 @@ module AppProfiler
       end
 
       def view(response: nil, autoredirect: nil, async: false)
-        id = AppProfiler::Viewer::RemoteViewer::SpeedscopeMiddleware.id(@profile.file)
+        id = Middleware.id(@profile.file)
 
         if response && response[0].to_i < 500
-          response[1]["Location"] = "/app_profiler/speedscope/viewer/#{id}"
+          response[1]["Location"] = "/app_profiler/firefox/viewer/#{id}"
           response[0] = 303
         else
           AppProfiler.logger.info("[Profiler] Profile available at /app_profiler/#{id}\n")
